update: imports and dependencies
