+++ crates/torii/core/Cargo.toml
@@ -15,8 +15,6 @@ async-trait.workspace = true
 chrono.workspace = true
 dojo-types = { path = "../../dojo-types" }
 dojo-world = { path = "../../dojo-world" }
-indexmap = "1.9.3"
-torii-graphql = { path = "../graphql" }
 
 log = "0.4.17"
 serde.workspace = true
@@ -28,6 +26,13 @@ tokio = { version = "1.20.1", features = [ "full" ] }
 tokio-stream = "0.1.11"
 tokio-util = "0.7.7"
 tracing.workspace = true
+
+#Dynamic subscriber
+async-stream = "0.3.0"
+futures-channel = "0.3.0"
+futures-util = "0.3.0"
+once_cell = "1.0"
+slab = "0.4.2"
 
 [dev-dependencies]
 camino.workspace = true
