+++ crates/torii/core/src/types.rs
@@ -1,6 +1,8 @@
 use core::fmt;
+use chrono::{DateTime, Utc};
 
 use serde::{Deserialize, Serialize};
+use sqlx::FromRow;
 use starknet::core::types::FieldElement;
 
 #[derive(Serialize, Deserialize)]
@@ -24,4 +26,24 @@ impl fmt::LowerHex for SQLFieldElement {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.0.fmt(f)
     }
+}
+
+#[derive(FromRow, Deserialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct Entity {
+    pub id: String,
+    pub keys: String,
+    pub component_names: String,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+}
+
+#[derive(FromRow, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct Component {
+    pub id: String,
+    pub name: String,
+    pub class_hash: String,
+    pub transaction_hash: String,
+    pub created_at: DateTime<Utc>,
 }
