+++ crates/torii/graphql/Cargo.toml
@@ -26,16 +26,10 @@ tokio-stream = "0.1.11"
 tokio-util = "0.7.7"
 tracing.workspace = true
 url = "2.2.2"
-#Dynamic subscriber
-async-stream = "0.3.0"
-futures-channel = "0.3.0"
-futures-util = "0.3.0"
-once_cell = "1.0"
-slab = "0.4.2"
+torii-core = { path = "../core" }
 
 [dev-dependencies]
 camino.workspace = true
 dojo-world = { path = "../../dojo-world" }
 starknet-crypto.workspace = true
 starknet.workspace = true
-torii-core = { path = "../core" }
