+++ crates/torii/graphql/src/lib.rs
@@ -4,7 +4,6 @@ pub mod object;
 mod query;
 pub mod schema;
 pub mod server;
-pub mod simple_broker;
 pub mod types;
 mod utils;
