+++ crates/torii/graphql/src/object/connection/page_info.rs
@@ -8,9 +8,9 @@ use crate::types::ScalarType;
 pub struct PageInfoObject {
     pub type_mapping: TypeMapping,
 }
-#[cfg_attr(feature = "cargo-clippy", allow(clippy::new_without_default))]
+
 impl PageInfoObject {
-    pub fn new() -> Self {
+    pub fn default() -> Self {
         Self {
             type_mapping: IndexMap::from([
                 (Name::new("hasPreviousPage"), TypeRef::named(TypeRef::BOOLEAN)),
